Refactor out strategy graph handling
